+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ dev
 ---
 - Fix bug in plot_pairs which crashes in case of 1 parameter
 - Fix bug in plot_marginals which outputs empty plots in case where we have parameter more than 5
-
+- Fix bug in plot_traces in BOLFI model
 0.7.2 (2018-06-20)
 ------------------
 - Added support for kwargs in elfi.set_client